audiostream-hd: Make framesSincePlay int64.

+++ framework/audiooutput-hd/AudioStreamHD.h
@@ -1,14 +1,16 @@
 #pragma once
 
+#include <stdint.h>
+
 class AudioStreamHD
 {
 public:
 	struct StreamInfo
 	{
-		float  frameRate;        // number of (single- or mult-channel) frames that are output per second
-		double secondsSincePlay; // the number of seconds since Play() was called
-		int    framesSincePlay;  // the number of frames since Play() was called
-		float  outputLatency;    // audio output latency in seconds. use this for precision timing when for instance scheduling notes
+		float   frameRate;        // number of (single- or mult-channel) frames that are output per second
+		double  secondsSincePlay; // the number of seconds since Play() was called
+		int64_t framesSincePlay;  // the number of frames since Play() was called
+		float   outputLatency;    // audio output latency in seconds. use this for precision timing when for instance scheduling notes
 	};
 	
 	struct ProvideInfo
